+++ tests/test_auth.py
@@ -3,18 +3,30 @@
 import pytest
 import respx
 
-from rtbhouse_sdk.client import BasicAuth, BasicTokenAuth, Client
+from rtbhouse_sdk.client import ApiTokenAuth, BasicAuth, BasicTokenAuth, Client
 
 
 @pytest.mark.parametrize(
     "auth_backend",
     (
+        ApiTokenAuth("token"),
         BasicAuth("user", "pwd"),
         BasicTokenAuth("token"),
     ),
 )
-def test_auth_backend_is_supported(auth_backend: BasicAuth | BasicTokenAuth) -> None:
+def test_auth_backend_is_supported(auth_backend: ApiTokenAuth | BasicAuth | BasicTokenAuth) -> None:
     Client(auth=auth_backend)
+
+
+def test_api_token_auth_flow(api_mock: respx.MockRouter) -> None:
+    api_mock.get("/example-endpoint").respond(200, json={"data": {}})
+
+    auth = ApiTokenAuth("abc")
+    with Client(auth) as cli:
+        cli._request("GET", "/example-endpoint")  # pylint: disable=protected-access
+
+    (call,) = api_mock.calls
+    assert call.request.headers["authorization"] == "Bearer abc"
 
 
 def test_basic_token_auth_flow(api_mock: respx.MockRouter) -> None:
@@ -22,7 +34,7 @@ def test_basic_token_auth_flow(api_mock: respx.MockRouter) -> None:
 
     auth = BasicTokenAuth("abc")
     with Client(auth=auth) as cli:
-        cli._get("/example-endpoint")  # pylint: disable=protected-access
+        cli._request("GET", "/example-endpoint")  # pylint: disable=protected-access
 
     (call,) = api_mock.calls
     assert call.request.headers["authorization"] == "Token abc"